+++ references.bib
@@ -6,7 +6,8 @@
   number={5},
   pages={1100--1111},
   year={2019},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acssynbio.9b00010}
 }
 
 @inproceedings{shah2018temporal,
@@ -15,7 +16,8 @@
   booktitle={International Conference on DNA Computing and Molecular Programming},
   pages={71--86},
   year={2018},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-00030-1_5}
 }
 
 @article{shah2019programming,
@@ -26,7 +28,8 @@
   number={4},
   pages={2668--2673},
   year={2019},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acs.nanolett.9b00590}
 }
 
 @article{wang2014localization,
@@ -37,5 +40,6 @@
   number={13},
   pages={15982--15991},
   year={2014},
-  publisher={Optical Society of America}
+  publisher={Optical Society of America},
+  doi = {10.1364/OE.22.015982}
 }
