### 2020-01-16 #2009 @shalinshah1993
Thank you for accepting the request! While not on that list, I used to be an active contributor to an open-source organization called SBML (https://github.com/nrnb) in the past and also know a few folks who are great software engineers and researchers, if they are interested in reviewing. https://github.com/draeger-lab https://github.com/niko-rodrigue
