### 2020-01-16 #2009 @csoneson
👋 @sirsharpest, @caldarolamartin, @lbugnon, @RealPolitiX - would you be interested in reviewing this submission for JOSS?

> ExTemp: A computational and image-processing suite for extracting temporal barcodes
