### 2020-01-16 #2009 @csoneson
@lbugnon Great, thanks! Reviewer instructions are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
Once we have at least two reviewers I will open the actual review issue, where the review will take place (you will have your reviewer checklist there).
