### 2020-01-17 #2026 @whedon
[REVIEW]: ExTemp: A computational and image-processing suite for extracting temporal barcodes
**Submitting author:** @shalinshah1993 (<a href="http://orcid.org/0000-0002-1406-3577">Shalin Shah</a>)
**Repository:** <a href="https://github.com/shalinshah1993/temporalDNAbarcodes" target ="_blank">https://github.com/shalinshah1993/temporalDNAbarcodes</a>
**Version:** 1.0.0
