### 2020-01-27 shalinshah1993/temporalDNAbarcodes#3 @SirSharpest
JOSS Reviewer comments:
Ref: [ExTemp Review](https://github.com/openjournals/joss-reviews/issues/2026)

# General comments

- Some minor grammar errors in the paper.pdf, e.g. "`In particular, since the fluorescent signal comes from dye-labeled DNA, hence temporal DNA barcodes`". To me this doesn't read-well
- The text itself seems a little light and hard to understand. I looked at the cited paper in Nano letters and it is explained much better. I raise this as a point in order to make any potential user more aware of what it is this software package does/works with.
- There are no acknowledgements, I see from your previous work that you are funded by `National Science Foundation Grants` (though, it is mentioned in the readme.MD).
- I don't find author contributions in the paper (though they are in the readme.MD).
- As you are using MATLAB, can you confirm what packages/toolboxes that a user will need, additionally, it would be useful to highlight which MATLAB versions the software should work on.
- I note you used on a HPC equiment to run your experiments, can you comment on if a 'standard' desktop PC is enough for this, or if HPC access is required to perform analysis in regular time?
- Your readme.MD mentions `they can be used for any arbitrary microscopy project that requires data extraction from fluorescence microscopy data` - perhaps expanding on this and showing some general usecases would be good to make this software useful for a wider audience?
- Have you tested in Octave, it would be useful to know if FOSS software is capable of running these modules rather than closed platforms.
- Do you provide estimations for expected errors from running your software, is there an alternative form of analysis? When similar experiments were ran prior to 2019, how were their data analysed?

# Review guidelines to be met:

- ` If there are any performance claims of the software, have they been confirmed` I cannot easily test this, are you able to provide any automated testing with example data
- The above comment also would address the requirements for: example usage, functionality and automated tests.
