### 2020-01-27 shalinshah1993/temporalDNAbarcodes#3 @AoifeHughes
I've re-read your readme.MD and I apologise, I missed the note you make about a cluster being required. This makes it difficult as a reviewer to confirm that the software runs and produces expected and valid outputs. Would you have suggestions of any easy methods to verify results?
