+++ paper.md
@@ -32,7 +32,7 @@ bibliography: references.bib
 
 # Summary
 
-We provide a package for fast and accurate extraction of temporal barcodes from a stack of microscopy images. This includes several steps such as object detection and tracking, 3D drift detection and correction, signal processing and denoising, and barcode extraction. A temporal barcode [@shah2019improved] is defined as the intensity trace over time of an object of interest. In particular, since the fluorescent signal comes from dye-labeled DNA, hence temporal DNA barcodes. To our knowledge, no other package can perform systematic extraction of time signals from microscopy images. The relevant application includes improved optical multiplexing and super-resolution imaging.
+We provide a package for fast and accurate extraction of temporal barcodes from a stack of microscopy images. This includes several steps such as object detection and tracking, 3D drift detection and correction, signal processing and denoising, and barcode extraction. A temporal barcode [@shah2019improved] is defined as the intensity trace over time of an object of interest. In particular, the fluorescent signal comes from a dye-labeled fluorescent DNA, hence it is referred to as *temporal DNA barcodes*. To our knowledge, no other package can perform systematic extraction of time signals from microscopy images. The relevant application includes improved optical multiplexing and super-resolution imaging.
 
 ``ExTemp`` is a set of MATLAB scripts that can act as computer vision pipeline to extract relevant information from the raw image stack. There are several steps involved to extract meaningful information from the recorded raw image stacks [@shah2019programming]. The first step includes data-collection \textit{i.e} recording a video (or an image stack) using a TIRF microscope. Once we have the raw data (usually several gigabytes), we convert the proprietary Leica lif file to a mat file using the bfmatlab library. This can help us with the development of the programmable downstream MATLAB scripts as the raw data is now available in the supported format. The next step includes the estimation and correction of the lateral (x-axis, y-axis) and axial (z-axis) drift. For lateral drift correction, we use the redundant cross-correlation algorithm [@wang2014localization] and for axial drift correction, we subtract mean pixel value of each frame.
 
