### 2020-01-27 shalinshah1993/temporalDNAbarcodes#3 @shalinshah1993
Thank you for the detailed review @SirSharpest

1. Done. Please check the paper. ` In particular, the fluorescent signal comes from a dye-labeled fluorescent DNA...`

2. Thank you! Yes, a lot of package details are also in the supplementary of the paper.

3. I've added them to the papers as well. Thanks for pointing this out!

4. I've added the author contributions to the paper as well. Thanks for pointing this out!

5. Thanks for pointing this out. We need Statistics and machine learning toolbox. I've added this in the dependencies section of README.md along with external dependencies. The version of MATLAB I used can be found in the system section however I believe it should work for anything >= 2018

6. Yes, we used an HPC system however you don't need it, in principle, to test. For example, if you run `lif2mat 'sample_data.lif'` it should convert proprietary Leica file to Matlab `.m` file as long as sample data is relatively small in size (a few hundred MBs). We require HPC since most large-scale microscopy experiments have raw data of several tens or hundreds of gigabytes.  (Refer to the end of this post to download the sample file from duke box)

7. We have not used or tested it in Octave. Unfortunately, it will be difficult at this point to have a dedicated person working on the Octave version of ExTemp. However, this will certainly be an exciting future alternative.

8. Maybe the supplementary section of the published paper can help? If so, we can cite it somewhere in the README. https://pubs.acs.org/doi/suppl/10.1021/acs.nanolett.9b00590/suppl_file/nl9b00590_si_001.pdf

Also, here is how I recommend you do an auto-test of the `sample_data.lif` (after downloading it from duke box https://duke.box.com/s/9xy67550rdawj6pk4woqxnyhwknq0n3c):

lif2mat('sample_data.lif')

This will process the lif file and produce Matlab .m file in /tmp/mat/ folder. Refer to README file https://github.com/shalinshah1993/temporalDNAbarcodes#module-description-sample-use to similarly try other portions of the package if you desire.
