### 2020-01-31 #2026 @csoneson
👋 @lbugnon, @SirSharpest, @RealPolitiX - I just wanted to check in on your reviews. Do you have an update on when you think you would be able to provide some comments for @shalinshah1993 (I see that @SirSharpest has opened an issue in the repo)?
