### 2020-01-31 #2026 @SirSharpest
> 👋 @lbugnon, @SirSharpest, @RealPolitiX - I just wanted to check in on your reviews. Do you have an update on when you think you would be able to provide some comments for @shalinshah1993 (I see that @SirSharpest has opened an issue in the repo)?

Yes, I'll have my review finished before Monday, I think that all my issues are addressed and just need to verify before I can sign off.
