### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @lbugnon
JOSS Reviewer comments: Demo not working
Ref openjournals/joss-reviews#2026
Hi, I'm trying to run the demo but it goes with the following error:

> Error using lif2mat (line 35)
> The size of the right hand side did not match the size of the indexing on the left hand side.

It appears that target frame_size is larger than framebycell.
