### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @shalinshah1993
What command did you type and what file did you run it on?
