### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @lbugnon
Sorry, I see that the avi file was deleted. Now I've tried sample_data.lif and it's working
