### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @lbugnon
Now lif is working, but findalllocalizations throws that 'filename' not in the path. Maybe I'm missing something (already addded this repo to the path)

> Error using VideoReader/init (line 601)
> The filename specified was not found in the MATLAB path.
>
> Error in VideoReader (line 171)
>             obj.init(fileName);
>
> Error in findalllocalizations (line 19)
>     videoObj = VideoReader(strcat('video/', file));
