### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @shalinshah1993
What command did you wrote?

Findalllocalizations would expect .m file in the right folder.

Regards,
Shalin

Sent from my iPhone

> On Jan 31, 2020, at 5:10 PM, Leandro A. Bugnon <notifications@github.com> wrote:
>
> ﻿
> Reopened #4.
>
