### 2020-01-31 shalinshah1993/temporalDNAbarcodes#4 @lbugnon
I'm trying to follow the readme pipeline. After lif2mat,
> findalllocalizations('sample_data.mat')

I tried moving the file from tmp/mat to  /mat but still same error

EDIT: now I see findalllocalizations is expecting a .avi, in the video/ folder, still not working with readme insttructions, please check the readme, I'm probably missing something:

> findalllocalizations('sample_data.avi')
> Not enough input arguments.
>
> Error in findalllocalizations (line 30)
>         cur_frame(cur_frame < threshold * mean(cur_frame)) = 0;
