### 2020-02-01 41fd2a7
Fixed issue for findalllocalization path
