### 2020-02-01 shalinshah1993/temporalDNAbarcodes#4 @shalinshah1993
Thank you for pointing this issue out, it was an error in README, I fixed it.  Basically, `findalllocalizations` expects a video file in the `video` folder.  I also added a sample video so you don't need to do any moving between folders yourself.

please do a git pull and then try running:

findalllocalizations('sample_data.avi')

This should find all the spots in the video file `sample_data.avi` found in video folder.
