### 2020-02-03 shalinshah1993/temporalDNAbarcodes#4 @lbugnon
Ok, some final issues to solve, mainly related with README and a proper testing scheme. Remember that having a manual or automatic testing procedure for the modules is a requisite in JOSS

- findalllocalizations and other functions  (as estdrift, crctdrift and calcdatatrend) expect other parameters besides filename. This should be stated in readme or set by default. Please check all the modules  so one can arrive to an expected outcome.

- When using findalllocalizations  with threashold=3 and 5, the following error appeears at some frames of sample_data
>Error using insertShape
>Expected POSITION to be finite

It seems to work ok when plotting is disabled.

- In the README,  "findlocalizations"appears in place of "gettemporalbarcode".

- Also the computer vision package is needed when toDisplayFrames is enabled.

- Wavelet package is required for denoising in gettemporalbarcode.

- There is some problem with R2018a when toDisplayFrames=1 (as you stated to use R2018b, this is not an issue, just saying)
