### 2020-02-03 4365f2f
fixed gettemporalbarcode readme
