### 2020-02-03 58f09ea
Fix for bug #6
