### 2020-02-03 shalinshah1993/temporalDNAbarcodes#4 @shalinshah1993
Thank you for raising these points. @lbugnon

Please see below:

- Thank you for pointing this issue. This has now been resolved with an updated README. Also, each file header has a short module description which states the expected parameter values.

- Thank you for pointing this. It seems I am not able to see this using this command:
`findalllocalizations(<file_name>, 3, 1, 1)`

   However, i do see it when I set a very high threshold:
   `findalllocalizations('ideal data  128x128.avi', 5, 1, 1)`

    This issue means that no point was detected in a frame. I've created a new issue and fixed it now. https://github.com/shalinshah1993/temporalDNAbarcodes/issues/6

- Thank you for pointing this out. The function use case in README has been fixed.

- Somehow MATLAB dependency library didn't state those as required. Thank you for letting us know. I have added both of them as required dependencies in the README (https://github.com/shalinshah1993/temporalDNAbarcodes#dependencies).

- Thank you for pointing this, I've created an issue referring to this. A fix suggesting that a MATLAB upgrade to >= R2018b is required.
