### 2020-02-03 shalinshah1993/temporalDNAbarcodes#3 @shalinshah1993
@SirSharpest if all your concerns are addressed, please close this issue. Thank you for your time and efforts with review :)
