### 2020-02-04 #2026 @csoneson
Thanks @lbugnon , @SirSharpest for your reviews.

@lbugnon - there are still some unchecked items in your checklist above. Could you confirm that these issues have been addressed, and if so, check the boxes?

@RealPolitiX - could you provide an update on your review status?
