### 2020-02-07 #2026 @RealPolitiX
Nice work, ready for publication
