### 2020-02-09 f1f5d23
Small edits to paper
