+++ references.bib
@@ -13,7 +13,7 @@
 @inproceedings{shah2018temporal,
   title={{Temporal DNA barcodes: A time-based approach for single-molecule imaging}},
   author={Shah, Shalin and Reif, John},
-  booktitle={International Conference on DNA Computing and Molecular Programming},
+  booktitle={{International Conference on DNA Computing and Molecular Programming}},
   pages={71--86},
   year={2018},
   organization={Springer},
@@ -23,7 +23,7 @@
 @article{shah2019programming,
   title={{Programming temporal DNA barcodes for single-molecule fingerprinting}},
   author={Shah, Shalin and Dubey, Abhishek K and Reif, John},
-  journal={Nano letters},
+  journal={Nano Letters},
   volume={19},
   number={4},
   pages={2668--2673},
@@ -35,7 +35,7 @@
 @article{wang2014localization,
   title={Localization events-based sample drift correction for localization microscopy with redundant cross-correlation algorithm},
   author={Wang, Yina and Schnitzbauer, Joerg and Hu, Zhe and Li, Xueming and Cheng, Yifan and Huang, Zhen-Li and Huang, Bo},
-  journal={Optics express},
+  journal={Optics Express},
   volume={22},
   number={13},
   pages={15982--15991},
