### 2020-02-09 #2026 @csoneson
👋 @shalinshah1993 - I have gone through your manuscript and sent a few suggestions in a PR. Could you please check, merge and generate a new proof here with `@whedon generate pdf`?

Next, please:
- [ ] check that the version information is consistently reported and report the version back here
- [ ] create a tagged release
- [ ] generate a zenodo archive
- [ ] make sure that the title and authors of the zenodo archive are identical to those of the paper (you can also add the ORCID of the authors there)
- [ ] report back here with the DOI of the zenodo archive

Finally, I would suggest to add also the JOSS paper to your README, and make it clear to users which of the papers to cite in which situation.
