### 2020-02-09 #2026 @shalinshah1993
Dear @csoneson you said you sent a few suggestions in PR. I'm sorry where did you send it?

- I created a release v1.0.0 https://github.com/shalinshah1993/temporalDNAbarcodes/releases/
- I have also enabled zendo integration for the repo. DOI: 10.5281/zenodo.3660411
- I have edited the README of the repo to make it more clear which papers to cite.
