### 2020-02-09 #2026 @shalinshah1993
Oh great, I accepted your PR. Thank you for those edits.

I think the authors' names and paper names should match now.
https://zenodo.org/record/3660411#.XkARhxNKjVo
