### 2020-02-09 #2026 @csoneson
@openjournals/joss-eics - I believe this paper is ready to be accepted!
