### 2020-02-09 #2026 @shalinshah1993
Thank you very much @csoneson for all your help and edits. Also, thank you @SirSharpest @lbugnon @RealPolitiX for your comments and feedback.

I think overall the `Extemp` package and GitHub repo look a lot more professional and clean. I'm happy that I submitted it to JOSS for peer-review.
