### 2020-02-09 #2026 @kthyng
Congratulations to @shalinshah1993 on your new publication! Thanks to @csoneson for editing and to @lbugnon, @SirSharpest, and @RealPolitiX for reviewing — we relied on your time and expertise!
